# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Movie API",
        "description": "API for managing films and screening",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/films": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Az összes filmet kilistázza",
                "operationId": "271e474415d2f29bf3c163247c5f4639",
                "responses": {
                    "200": {
                        "description": "Az összes filmet kilistázza"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/films/show/{id}": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Get a specific film by ID",
                "operationId": "8ad2497b89a853c96b7c561efdc14c2a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the film to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Film details"
                    },
                    "404": {
                        "description": "Film not found"
                    }
                }
            }
        },
        "/api/films/createNewFilm": {
            "post": {
                "tags": [
                    "Films"
                ],
                "summary": "Create a new film",
                "description": "Creates a new film with the provided details.",
                "operationId": "createFilm",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "age_rating",
                                    "language"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Inception"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A mind-bending thriller."
                                    },
                                    "age_rating": {
                                        "type": "string",
                                        "example": "PG-13"
                                    },
                                    "language": {
                                        "type": "string",
                                        "example": "English"
                                    },
                                    "cover_image": {
                                        "type": "string",
                                        "example": "https://example.com/image.jpg",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Film created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Inception"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "A mind-bending thriller."
                                        },
                                        "age_rating": {
                                            "type": "string",
                                            "example": "PG-13"
                                        },
                                        "language": {
                                            "type": "string",
                                            "example": "English"
                                        },
                                        "cover_image": {
                                            "type": "string",
                                            "example": "https://example.com/image.jpg"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2025-03-26T18:56:49.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2025-03-26T18:56:49.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/films/update/{id}": {
            "put": {
                "tags": [
                    "Films"
                ],
                "summary": "Update a film by ID",
                "description": "Update the details of a specific film by its ID",
                "operationId": "016f9b24cca70cfd1b6c1d33a7b10e6a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the film to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "age_rating",
                                    "language"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "The Matrix"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A computer hacker learns from mysterious rebels about the true nature of his reality and his role in the war against its controllers."
                                    },
                                    "age_rating": {
                                        "type": "integer",
                                        "example": 18
                                    },
                                    "language": {
                                        "type": "string",
                                        "example": "English"
                                    },
                                    "cover_image": {
                                        "type": "string",
                                        "example": "matrix_cover.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Film successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Film"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Film not found"
                    }
                }
            }
        },
        "/api/films/delete/{id}": {
            "delete": {
                "tags": [
                    "Films"
                ],
                "summary": "Delete a film by ID. Delete the Screen as well",
                "description": "This endpoint deletes a film from the database by its ID.",
                "operationId": "deleteFilm",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the film to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Film deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Film deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Film not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Film not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/films/search": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Search films by title, language, and age rating",
                "description": "Returns a list of films that match the search criteria. Filters can be applied based on title, language, and age rating.",
                "operationId": "searchFilms",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title of the film to search for",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Matrix"
                        }
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "description": "Language of the film",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "English"
                        }
                    },
                    {
                        "name": "age_rating",
                        "in": "query",
                        "description": "Age rating of the film",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "12"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of films that match the search criteria",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Film"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid search parameters"
                    },
                    "404": {
                        "description": "No films found matching the criteria"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Film": {
                "required": [
                    "id",
                    "title",
                    "description",
                    "age_rating",
                    "language",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "The Matrix"
                    },
                    "description": {
                        "type": "string",
                        "example": "A computer hacker learns from mysterious rebels about the true nature of his reality and his role in the war against its controllers."
                    },
                    "age_rating": {
                        "type": "integer",
                        "example": 18
                    },
                    "language": {
                        "type": "string",
                        "example": "English"
                    },
                    "cover_image": {
                        "type": "string",
                        "example": "matrix_cover.jpg"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-26T18:56:49"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-26T18:56:49"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}